run.shを修正

+++ run.sh
@@ -1,10 +1,18 @@
+#!/bin/bash
 make
 
-mkdir models
+if [ ! -d "models" ]; then
+  mkdir models
+fi
+
 cd models
-wget http://graphics.stanford.edu/pub/3Dscanrep/bunny.tar.gz
-tar -zxvf bunny.tar.gz
-rm bunny.tar.gz
+
+if [ ! -d "bunny" ]; then
+  wget http://graphics.stanford.edu/pub/3Dscanrep/bunny.tar.gz
+  tar -zxvf bunny.tar.gz
+  rm bunny.tar.gz
+fi
+
 cd ..
 
 ./main -n 3 -w 320 -h 240 -s 2 -ss 8
